Agent, Player and Organize update !
Player is now a class who represent the user in the game, it contains all the old parameters who were necessary for the game to run.

Agent is now a class who extends Player, this class is meant for the DeepLearning agent to operate and win the game by himself.

Menu got additional square meant fot the Agent activation and now the little colored squares will not show up the OptionTiles.

Main got a lot of changes to ensure a more readable and efficient code.

Some more improvements and changes.